reduce list of currencies to a subset
